+++ cinema_expert_database/cinema_expert_movies_genres.sql
@@ -23,8 +23,8 @@ DROP TABLE IF EXISTS `movies_genres`;
 /*!40101 SET @saved_cs_client     = @@character_set_client */;
 /*!50503 SET character_set_client = utf8mb4 */;
 CREATE TABLE `movies_genres` (
-  `movie_id` int NOT NULL,
-  `genre_id` int NOT NULL,
+  `movie_id` bigint NOT NULL,
+  `genre_id` bigint NOT NULL,
   PRIMARY KEY (`movie_id`,`genre_id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;
 /*!40101 SET character_set_client = @saved_cs_client */;
